[FOLD] Adopt changes suggested by @jwbusch in review

+++ src/ripple/overlay/impl/Message.cpp
@@ -39,8 +39,7 @@ Message::Message (::google::protobuf::Message const& message, int type)
         message.SerializeToArray (&mBuffer [Message::kHeaderBytes], messageBytes);
     }
 
-    mCategory = safe_cast<int>(TrafficCount::categorize
-        (message, type, false));
+    mCategory = TrafficCount::categorize(message, type, false);
 }
 
 bool Message::operator== (Message const& other) const
